feat(core): add source map annotations to nx show target
- Show where each config property comes from (file + plugin) in both
  JSON and text output via source map integration
- Reorganize output order: command/executor first, then small metadata
  (parallelism, continuous, cache), then larger blocks (inputs, outputs,
  options, configurations)
- Smart command/executor hoisting: show Command for run-commands with a
  single command, Script for run-script targets, Executor otherwise
- Pretty-print complex option values with indented JSON
- Track input provenance through named input expansion so each expanded
  input line shows its source
- Refactor into vertical slices under show-target/ directory: each
  subcommand (info, inputs, outputs) owns its handler, data resolution,
  and rendering in a single file
- Split monolith test file into per-slice specs with shared test-utils
