+++ packages/nx/src/command-line/show/show-target/index.ts
@@ -1,0 +1,3 @@
+export { showTargetInfoHandler } from './info';
+export { showTargetInputsHandler } from './inputs';
+export { showTargetOutputsHandler } from './outputs';
