+++ packages/nx/src/command-line/show/show-target/info.spec.ts
@@ -1,0 +1,463 @@
+import {
+  GraphBuilder,
+  setupBeforeEach,
+  setupAfterEach,
+  setGraph,
+  setMockSourceMaps,
+} from './test-utils';
+import { showTargetInfoHandler } from './info';
+
+describe('show target info', () => {
+  beforeEach(setupBeforeEach);
+  afterEach(setupAfterEach);
+
+  it('should show target info when project:target is provided', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                options: { outputPath: 'dist/apps/my-app' },
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.project).toBe('my-app');
+    expect(parsed.target).toBe('build');
+    expect(parsed.executor).toBe('@nx/web:build');
+    expect(parsed.options).toEqual({ outputPath: 'dist/apps/my-app' });
+  });
+
+  it('should infer project from cwd when only target name given', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: { build: { executor: '@nx/web:build' } },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    process.cwd = jest.fn().mockReturnValue('/workspace/apps/my-app');
+
+    await showTargetInfoHandler({ target: 'build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.project).toBe('my-app');
+    expect(parsed.target).toBe('build');
+  });
+
+  it('should output valid JSON with expected structure', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                options: { outputPath: 'dist/apps/my-app' },
+                outputs: ['{options.outputPath}'],
+                inputs: ['default', '^default'],
+                cache: true,
+                configurations: { production: { optimization: true } },
+                defaultConfiguration: 'production',
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed).toMatchObject({
+      project: 'my-app',
+      target: 'build',
+      executor: '@nx/web:build',
+      options: { outputPath: 'dist/apps/my-app' },
+      inputs: [{ fileset: '{projectRoot}/**/*' }, '^default'],
+      outputs: ['{options.outputPath}'],
+      configurations: ['production'],
+      defaultConfiguration: 'production',
+      cache: true,
+      parallelism: true,
+    });
+  });
+
+  it('should merge configuration options correctly', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                options: {
+                  outputPath: 'dist/apps/my-app',
+                  optimization: false,
+                },
+                configurations: {
+                  production: { optimization: true, sourceMap: false },
+                },
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await showTargetInfoHandler({
+      target: 'my-app:build',
+      configuration: 'production',
+      json: true,
+    });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.options).toEqual({
+      outputPath: 'dist/apps/my-app',
+      optimization: true,
+      sourceMap: false,
+    });
+  });
+
+  it('should expand dependsOn ^build to flat taskId list', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: { executor: '@nx/web:build', dependsOn: ['^build'] },
+            },
+          },
+          'app'
+        )
+        .addProjectConfiguration(
+          {
+            root: 'libs/lib-a',
+            name: 'lib-a',
+            targets: { build: { executor: '@nx/js:tsc' } },
+          },
+          'lib'
+        )
+        .addProjectConfiguration(
+          {
+            root: 'libs/lib-b',
+            name: 'lib-b',
+            targets: { build: { executor: '@nx/js:tsc' } },
+          },
+          'lib'
+        )
+        .addDependency('my-app', 'lib-a')
+        .addDependency('my-app', 'lib-b')
+        .build()
+    );
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.dependsOn.sort()).toEqual(['lib-a:build', 'lib-b:build']);
+  });
+
+  it('should expand self-reference dependsOn to same project taskId', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: { executor: '@nx/web:build' },
+              test: { executor: '@nx/jest:jest', dependsOn: ['build'] },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await showTargetInfoHandler({ target: 'my-app:test', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.dependsOn).toContain('my-app:build');
+  });
+
+  it('should error when target not found and list available targets', async () => {
+    const { output } = require('../../../utils/output');
+    jest.spyOn(process, 'exit').mockImplementation((code) => {
+      throw new Error(`process.exit: ${code}`);
+    });
+
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: { executor: '@nx/web:build' },
+              test: { executor: '@nx/jest:jest' },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await expect(
+      showTargetInfoHandler({ target: 'my-app:serve', json: true })
+    ).rejects.toThrow('process.exit: 1');
+
+    expect(output.error).toHaveBeenCalledWith(
+      expect.objectContaining({
+        title: expect.stringContaining('serve'),
+        bodyLines: expect.arrayContaining([expect.stringContaining('build')]),
+      })
+    );
+  });
+
+  it('should error when project not found', async () => {
+    const { output } = require('../../../utils/output');
+    jest.spyOn(process, 'exit').mockImplementation((code) => {
+      throw new Error(`process.exit: ${code}`);
+    });
+
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: { build: { executor: '@nx/web:build' } },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await expect(
+      showTargetInfoHandler({ target: 'nonexistent:build', json: true })
+    ).rejects.toThrow('process.exit: 1');
+
+    expect(output.error).toHaveBeenCalledWith(
+      expect.objectContaining({
+        title: expect.stringContaining('nonexistent'),
+      })
+    );
+  });
+
+  it('should error when configuration not found and list available configs', async () => {
+    const { output } = require('../../../utils/output');
+    jest.spyOn(process, 'exit').mockImplementation((code) => {
+      throw new Error(`process.exit: ${code}`);
+    });
+
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                configurations: { production: {}, staging: {} },
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    await expect(
+      showTargetInfoHandler({
+        target: 'my-app:build',
+        configuration: 'nonexistent',
+        json: true,
+      })
+    ).rejects.toThrow('process.exit: 1');
+
+    expect(output.error).toHaveBeenCalledWith(
+      expect.objectContaining({
+        title: expect.stringContaining('nonexistent'),
+        bodyLines: expect.arrayContaining([
+          expect.stringContaining('production'),
+        ]),
+      })
+    );
+  });
+
+  it('should include sourceMap in JSON output when source maps are available', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                options: { outputPath: 'dist/apps/my-app' },
+                cache: true,
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    setMockSourceMaps({
+      'apps/my-app': {
+        'targets.build': ['apps/my-app/project.json', '@nx/web'],
+        'targets.build.executor': ['apps/my-app/project.json', '@nx/web'],
+        'targets.build.options.outputPath': [
+          'apps/my-app/project.json',
+          '@nx/web',
+        ],
+        'targets.build.cache': ['nx.json', 'nx/core/project-json'],
+      },
+    });
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.sourceMap).toBeDefined();
+    expect(parsed.sourceMap.executor).toEqual([
+      'apps/my-app/project.json',
+      '@nx/web',
+    ]);
+    expect(parsed.sourceMap['options.outputPath']).toEqual([
+      'apps/my-app/project.json',
+      '@nx/web',
+    ]);
+    expect(parsed.sourceMap.cache).toEqual(['nx.json', 'nx/core/project-json']);
+    expect(parsed.sourceMap.target).toEqual([
+      'apps/my-app/project.json',
+      '@nx/web',
+    ]);
+  });
+
+  it('should not include sourceMap when source maps are empty', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: { build: { executor: '@nx/web:build' } },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    setMockSourceMaps({});
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.sourceMap).toBeUndefined();
+  });
+
+  it('should show source annotations in text output', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: {
+                executor: '@nx/web:build',
+                options: { outputPath: 'dist/apps/my-app' },
+              },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    setMockSourceMaps({
+      'apps/my-app': {
+        'targets.build': ['apps/my-app/project.json', '@nx/web'],
+        'targets.build.executor': ['apps/my-app/project.json', '@nx/web'],
+        'targets.build.options.outputPath': [
+          'apps/my-app/project.json',
+          '@nx/web',
+        ],
+      },
+    });
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: false });
+
+    const allLogged = (console.log as jest.Mock).mock.calls
+      .map((c) => c[0])
+      .join('\n');
+    expect(allLogged).toContain('apps/my-app/project.json');
+    expect(allLogged).toContain('@nx/web');
+  });
+
+  it('should only include source map entries for the requested target', async () => {
+    setGraph(
+      new GraphBuilder()
+        .addProjectConfiguration(
+          {
+            root: 'apps/my-app',
+            name: 'my-app',
+            targets: {
+              build: { executor: '@nx/web:build' },
+              test: { executor: '@nx/jest:jest' },
+            },
+          },
+          'app'
+        )
+        .build()
+    );
+
+    setMockSourceMaps({
+      'apps/my-app': {
+        'targets.build': ['apps/my-app/project.json', '@nx/web'],
+        'targets.build.executor': ['apps/my-app/project.json', '@nx/web'],
+        'targets.test': ['apps/my-app/project.json', '@nx/jest'],
+        'targets.test.executor': ['apps/my-app/project.json', '@nx/jest'],
+      },
+    });
+
+    await showTargetInfoHandler({ target: 'my-app:build', json: true });
+
+    const parsed = JSON.parse((console.log as jest.Mock).mock.calls[0][0]);
+    expect(parsed.sourceMap).toBeDefined();
+    expect(parsed.sourceMap.executor).toEqual([
+      'apps/my-app/project.json',
+      '@nx/web',
+    ]);
+    expect(parsed.sourceMap['targets.test']).toBeUndefined();
+    expect(parsed.sourceMap['targets.test.executor']).toBeUndefined();
+  });
+});
